PR847/Indan: updateTimestamp before taking kernel lock
The intention of the MCS design is that updateTimestamp is performed
as early in the kernel operation instruction stream as possible.
Acquiring the lock can take approximately WCET time.

For timer interrupts this is needed to prevent periodic tasks that
run out of budget being postponed by low priority tasks, effectively
increasing their period with the average lock take time. This will
cause periodic tasks to drift in regard to the system clock and each
other, making the system less deterministic.

See also seL4 issue #844.

Signed-off-by: Indan Zupancic <Indan.Zupancic@mep-info.com>