boot: create_untypeds_for_phys_region()
Creating the untypeds is the place in the boot process where the switch
from plain physical addresses to addressing physical memory via the
kernel window happens. Passing physical region creates a single place of
control, to check what can be addressed and what not.
The region (or a part of) it could be outside the kernel window, so it
cannot be accessed. Ensure to never generate a pptr outside of that
window.

Signed-off-by: Axel Heider <axelheider@gmx.de>